Supply and Demand of Boston Airbnb's
In this project, we will empirically examine the fundamental concepts of supply and demand through Airbnb rentals in Boston. We will also analyze the effects of occupancy taxes on the short-term rental market, which were enacted by the Massachusetts state legislature on July 1st, 2019.

The project is roughly divided into 3 large parts:

In the first part, you will conduct some data processing and create a demand curve from (imputed) Airbnb transactions.
In the second part, you will create a supply curve for Airbnb listings and determine the market equilibrium.
In the third part, you will examine the effects of the occupancy tax on the market.